Implemented damage calculations
Note, this does not round properly.
In the first four generations, during the calculation, all operations are carried out on integers
internally—this means that all division operations are truncated integer division
(i.e. rounding down if the result is not an integer), and the results of multiplication operations
are rounded down afterwards (truncating any fractional part). From Generation V onward,
 there are three different types of rounding; a flooring (the same as previous generations),
rounding to the nearest integer while rounding down at 0.5, and rounding to the nearest integer
while rounding up at 0.5.